# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UAE Freelance Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /freelancers:
    get:
      operationId: listFreelancers
      tags: [users]
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: skill
          schema: { type: string }
        - in: query
          name: minRate
          schema: { type: number }
        - in: query
          name: maxRate
          schema: { type: number }
      responses:
        "200":
          description: Freelancers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreelancerCard"
  /freelancers/{id}:
    get:
      operationId: getFreelancer
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerDetail"
  /profiles/freelancer:
    put:
      operationId: updateFreelancerProfile
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFreelancerProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerDetail"
  /profiles/client:
    put:
      operationId: updateClientProfile
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
...
components:
  schemas:
...
    FreelancerCard:
      type: object
      required: [id, fullName, headline, skills, hourlyRate, currency, ratingAvg, ratingCount, location]
      properties:
        id: { type: integer }
        fullName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        headline: { type: string }
        skills:
          type: array
          items: { type: string }
        hourlyRate: { type: number }
        currency: { type: string }
        ratingAvg: { type: number }
        ratingCount: { type: integer }
        location: { type: string }
    FreelancerDetail:
      allOf:
        - $ref: "#/components/schemas/FreelancerCard"
        - type: object
          required: [bio, portfolio, completedJobs]
          properties:
            bio: { type: string }
            portfolio:
              type: array
              items:
                type: object
                required: [title, url]
                properties:
                  title: { type: string }
                  url: { type: string }
                  description: { type: ["string", "null"] }
            completedJobs: { type: integer }
    UpdateFreelancerProfileBody:
      type: object
      properties:
        fullName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        headline: { type: string }
        bio: { type: string }
        skills:
          type: array
          items: { type: string }
        hourlyRate: { type: number }
        location: { type: string }
        portfolio:
          type: array
          items:
            type: object
            required: [title, url]
            properties:
              title: { type: string }
              url: { type: string }
              description: { type: ["string", "null"] }
    ClientDetail:
      type: object
      required: [id, fullName, companyName, overview, activeJobs, completedJobs, ratingAvg, ratingCount]
      properties:
        id: { type: integer }
        fullName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        companyName: { type: string }
        logoUrl: { type: ["string", "null"] }
        overview: { type: string }
        location: { type: string }
        activeJobs: { type: integer }
        completedJobs: { type: integer }
        ratingAvg: { type: number }
        ratingCount: { type: integer }
    UpdateClientProfileBody:
      type: object
      properties:
        fullName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        companyName: { type: string }
        logoUrl: { type: ["string", "null"] }
        overview: { type: string }
        location: { type: string }
...